requirements.txt
flask
flask-cors
python-dotenv
stripe
pyjwt
werkzeug
